change from int to float

+++ commands/mul.js
@@ -11,8 +11,8 @@ module.exports = {
 		// const now = Date.now();
 		// const duration = (now - msg.createdTimestamp) / 1000;
 		// msg.reply(`Pong! Time needed: ${duration.toFixed(3)}s`);
-        var a = parseInt(args[0]);
-        var b = parseInt(args[1]);
+        var a = parseFloat(args[0]);
+        var b = parseFloat(args[1]);
         msg.reply(`Multiplication of two numbers is ${a * b}`);
 	},
 };
